+++ src/pages/not-found-page.module.scss
@@ -1,0 +1,13 @@
+.page {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+
+  &Link {
+    color: #D0BCFF;
+    text-decoration: none;
+  }
+}
