+++ src/shared/ui/form/form.module.scss
@@ -14,6 +14,15 @@
     background-color: #D0BCFF;
     border: none;
     box-shadow: 2px 2px 2px rgb(234 221 255 / 30%);
+    transition: box-shadow 0.2s;
+
+    &:disabled {
+      background-color: #E6E0E9;
+    }
+
+    &:hover {
+      box-shadow: 2px 2px 4px 2px rgb(234 221 255 / 30%);
+    }
 
     &:not(:disabled) {
       cursor: pointer;
