+++ src/shared/ui/layout.module.scss
@@ -15,4 +15,5 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  height: 100%;
 }
